Add resource base authorization to the CoursesController.

+++ PortalTeme/API/Controllers/CoursesController.cs
@@ -13,24 +13,38 @@ using System.Threading.Tasks;
 namespace PortalTeme.API.Controllers {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Policy = AuthorizationConstants.AdministratorPolicy)]
+    [Authorize]
     public class CoursesController : ControllerBase {
         private readonly PortalTemeContext _context;
+        private readonly IAuthorizationService authorizationService;
 
-        public CoursesController(PortalTemeContext context) {
+        public CoursesController(PortalTemeContext context, IAuthorizationService authorizationService) {
             _context = context;
+            this.authorizationService = authorizationService;
         }
 
         // GET: api/Courses
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Course>>> GetCourses() {
-            return await _context.Courses.ToListAsync();
+
+            var courses = await _context.Courses.ToListAsync();
+            var results = new List<Course>();
+            foreach (var course in courses) {
+                if ((await authorizationService.AuthorizeAsync(User, course, AuthorizationConstants.CanViewCoursesPolicy)).Succeeded)
+                    results.Add(course);
+            }
+
+            return results;
         }
 
         // GET: api/Courses/5
         [HttpGet("{id}")]
         public async Task<ActionResult<Course>> GetCourse(Guid id) {
             var course = await _context.Courses.FindAsync(id);
+
+            var authorization = await authorizationService.AuthorizeAsync(User, course, AuthorizationConstants.CanViewCoursesPolicy);
+            if (!authorization.Succeeded)
+                return Forbid();
 
             if (course is null)
                 return NotFound();
@@ -40,9 +54,14 @@ namespace PortalTeme.API.Controllers {
 
         // PUT: api/Courses/5
         [HttpPut("{id}")]
+        [Authorize(Policy = AuthorizationConstants.CanUpdateCoursePolicy)]
         public async Task<IActionResult> PutCourse(Guid id, Course course) {
             if (id != course.Id)
                 return BadRequest();
+
+            var authorization = await authorizationService.AuthorizeAsync(User, course, AuthorizationConstants.CanUpdateCoursePolicy);
+            if (!authorization.Succeeded)
+                return Forbid();
 
             _context.Entry(course).State = EntityState.Modified;
 
@@ -60,7 +79,13 @@ namespace PortalTeme.API.Controllers {
 
         // POST: api/Courses
         [HttpPost]
+        [Authorize(Policy = AuthorizationConstants.CanUpdateCoursePolicy)]
         public async Task<ActionResult<Course>> PostCourse(Course course) {
+
+            var authorization = await authorizationService.AuthorizeAsync(User, AuthorizationConstants.CanCreateCoursePolicy);
+            if (!authorization.Succeeded)
+                return Forbid();
+
             _context.Courses.Add(course);
             await _context.SaveChangesAsync();
 
@@ -69,10 +94,15 @@ namespace PortalTeme.API.Controllers {
 
         // DELETE: api/Courses/5
         [HttpDelete("{id}")]
+        [Authorize(Policy = AuthorizationConstants.CanDeleteCoursePolicy)]
         public async Task<ActionResult<Course>> DeleteCourse(Guid id) {
             var course = await _context.Courses.FindAsync(id);
             if (course is null)
                 return NotFound();
+
+            var authorization = await authorizationService.AuthorizeAsync(User, course, AuthorizationConstants.CanDeleteCoursePolicy);
+            if (!authorization.Succeeded)
+                return Forbid();
 
             _context.Courses.Remove(course);
             await _context.SaveChangesAsync();
